Move controlbar.
